// MAVI.Client/src/app/models/Post.ts
export interface Post {
    userPhoto: string;       
    username: string;        
    time: string;        
    photo: string;         
    likesPhotos: string[];  
    likedBy: string;         
    likesCount: number;       
    caption: string;         
    tag: string;     
    commentsText: string;
  }
  